@@ -3,99 +3,98 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@babel/core": "7.12.3",
-    "@babel/node": "7.13.13",
-    "@pmmmwh/react-refresh-webpack-plugin": "0.4.3",
-    "@svgr/webpack": "5.5.0",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-relay": "11.0.0",
-    "@typescript-eslint/eslint-plugin": "^4.5.0",
-    "@typescript-eslint/parser": "^4.5.0",
-    "babel-eslint": "^10.1.0",
-    "babel-jest": "^26.6.0",
-    "babel-loader": "8.1.0",
-    "babel-plugin-named-asset-import": "^0.3.7",
-    "babel-plugin-relay": "11.0.1",
-    "babel-preset-react-app": "^10.0.0",
+    "@babel/core": "^7.16.0",
+    "@babel/node": "^7.16.8",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.3",
+    "@svgr/webpack": "^5.5.0",
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.4.0",
+    "@types/node": "^16.11.22",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-relay": "^13.0.1",
+    "babel-jest": "^27.4.2",
+    "babel-loader": "^8.2.3",
+    "babel-plugin-named-asset-import": "^0.3.8",
+    "babel-plugin-relay": "^13.0.2",
+    "babel-preset-react-app": "^10.0.1",
     "bfj": "^7.0.2",
-    "camelcase": "^6.1.0",
-    "case-sensitive-paths-webpack-plugin": "2.3.0",
-    "classnames": "2.3.1",
-    "concurrently": "6.0.0",
-    "css-loader": "4.3.0",
-    "dotenv": "8.2.0",
-    "dotenv-expand": "5.1.0",
-    "eslint": "^7.11.0",
-    "eslint-config-react-app": "^6.0.0",
-    "eslint-plugin-flowtype": "^5.2.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest": "^24.1.0",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-react": "^7.21.5",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-testing-library": "^3.9.2",
-    "eslint-webpack-plugin": "^2.5.2",
-    "express": "4.17.1",
-    "express-graphql": "0.12.0",
-    "file-loader": "6.1.1",
-    "fs-extra": "^9.0.1",
-    "get-graphql-schema": "2.1.2",
-    "graphql": "15.5.0",
-    "graphql-relay": "0.6.0",
-    "html-webpack-plugin": "4.5.0",
-    "identity-obj-proxy": "3.0.0",
-    "jest": "26.6.0",
-    "jest-circus": "26.6.0",
-    "jest-resolve": "26.6.0",
-    "jest-watch-typeahead": "0.6.1",
-    "mini-css-extract-plugin": "0.11.3",
-    "npm-watch": "0.9.0",
-    "optimize-css-assets-webpack-plugin": "5.0.4",
-    "pnp-webpack-plugin": "1.6.4",
-    "postcss-flexbugs-fixes": "4.2.1",
-    "postcss-loader": "3.0.0",
-    "postcss-normalize": "8.0.1",
-    "postcss-preset-env": "6.7.0",
-    "postcss-safe-parser": "5.0.2",
-    "prompts": "2.4.0",
+    "browserslist": "^4.18.1",
+    "camelcase": "^6.2.1",
+    "case-sensitive-paths-webpack-plugin": "^2.4.0",
+    "classnames": "^2.3.1",
+    "concurrently": "^7.0.0",
+    "css-loader": "^6.5.1",
+    "css-minimizer-webpack-plugin": "^3.2.0",
+    "dotenv": "^10.0.0",
+    "dotenv-expand": "^5.1.0",
+    "eslint": "^8.3.0",
+    "eslint-config-react-app": "^7.0.0",
+    "eslint-webpack-plugin": "^3.1.1",
+    "express": "^4.17.2",
+    "express-graphql": "^0.12.0",
+    "file-loader": "^6.2.0",
+    "fs-extra": "^10.0.0",
+    "graphql": "^15.3.0",
+    "graphql-relay": "^0.7.0",
+    "html-webpack-plugin": "^5.5.0",
+    "husky": "^7.0.4",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^27.4.3",
+    "jest-resolve": "^27.4.2",
+    "jest-watch-typeahead": "^1.0.0",
+    "lint-staged": "^12.3.3",
+    "mini-css-extract-plugin": "^2.4.5",
+    "npm-watch": "^0.11.0",
+    "postcss": "^8.4.4",
+    "postcss-flexbugs-fixes": "^5.0.2",
+    "postcss-loader": "^6.2.1",
+    "postcss-normalize": "^10.0.1",
+    "postcss-preset-env": "^7.0.1",
+    "prettier": "^2.5.1",
+    "prompts": "^2.4.2",
     "react": "^17.0.2",
-    "react-app-polyfill": "^2.0.0",
-    "react-dev-utils": "^11.0.3",
+    "react-app-polyfill": "^3.0.0",
+    "react-dev-utils": "^12.0.0",
     "react-dom": "^17.0.2",
-    "react-error-boundary": "3.1.1",
-    "react-refresh": "^0.8.3",
-    "react-relay": "11.0.1",
-    "relay-compiler": "11.0.1",
-    "relay-compiler-language-typescript": "13.0.6",
-    "resolve": "1.18.1",
-    "resolve-url-loader": "^3.1.2",
-    "sass-loader": "^10.0.5",
-    "semver": "7.3.2",
-    "style-loader": "1.3.0",
-    "terser-webpack-plugin": "4.2.3",
-    "todomvc-app-css": "2.3.0",
-    "ts-pnp": "1.2.0",
-    "typescript": "4.2.3",
-    "url-loader": "4.1.1",
-    "web-vitals": "^1.0.1",
-    "webpack": "4.44.2",
-    "webpack-dev-server": "3.11.1",
-    "webpack-manifest-plugin": "2.2.0",
-    "workbox-webpack-plugin": "5.1.4"
+    "react-error-boundary": "^3.1.4",
+    "react-refresh": "^0.11.0",
+    "react-relay": "^13.0.2",
+    "relay-compiler": "^13.0.2",
+    "resolve": "^1.20.0",
+    "resolve-url-loader": "^4.0.0",
+    "sass-loader": "^12.3.0",
+    "semver": "^7.3.5",
+    "source-map-loader": "^3.0.0",
+    "style-loader": "^3.3.1",
+    "terser-webpack-plugin": "^5.2.5",
+    "typescript": "^4.5.5",
+    "web-vitals": "^2.1.4",
+    "webpack": "^5.64.4",
+    "webpack-dev-server": "^4.6.0",
+    "webpack-manifest-plugin": "^4.0.2",
+    "workbox-webpack-plugin": "^6.4.1"
+  },
+  "lint-staged": {
+    "src/**/*.{js,jsx,ts,tsx,json,css,scss,md}": [
+      "prettier --write"
+    ]
   },
   "scripts": {
     "start": "concurrently --kill-others --names \"server,compiler\" \"NODE_ENV=development babel-node scripts/start.js\" \"yarn run watch_relay\"",
     "build": "node scripts/build.js",
     "test": "node scripts/test.js",
     "update-schema": "NODE_ENV=development babel-node scripts/update-schema.js",
-    "relay": "relay-compiler --src ./src --schema schema/schema.graphql --language typescript --artifactDirectory ./src/__generated__",
+    "relay": "relay-compiler $@",
     "watch_relay": "npm-watch relay"
+  },
+  "relay": {
+    "src": "src",
+    "schema": "schema/schema.graphql",
+    "artifactDirectory": "src/__generated__",
+    "language": "typescript"
   },
   "watch": {
     "relay": {
