@@ -1,6 +1,6 @@
 'use strict';
 
-const babelJest = require('babel-jest');
+const babelJest = require('babel-jest').default;
 
 const hasJsxRuntime = (() => {
   if (process.env.DISABLE_NEW_JSX_TRANSFORM === 'true') {
