@@ -1,31 +1,35 @@
+/**
+ * @generated SignedSource<<f09fc7dbc5938e42f44385965c29d77e>>
+ * @lightSyntaxTransform
+ * @nogrep
+ */
+
 /* tslint:disable */
 /* eslint-disable */
 // @ts-nocheck
 
-import { ReaderFragment } from "relay-runtime";
+import { Fragment, ReaderFragment } from 'relay-runtime';
 import { FragmentRefs } from "relay-runtime";
-export type TodoListFooter_user = {
-    readonly id: string;
-    readonly userId: string;
-    readonly completedCount: number;
-    readonly todos: {
-        readonly edges: ReadonlyArray<{
-            readonly node: {
-                readonly id: string;
-                readonly complete: boolean;
-            } | null;
-        } | null> | null;
-    } | null;
-    readonly totalCount: number;
-    readonly " $refType": "TodoListFooter_user";
+export type TodoListFooter_user$data = {
+  readonly id: string;
+  readonly userId: string;
+  readonly completedCount: number;
+  readonly todos: {
+    readonly edges: ReadonlyArray<{
+      readonly node: {
+        readonly id: string;
+        readonly complete: boolean;
+      } | null;
+    } | null> | null;
+  } | null;
+  readonly totalCount: number;
+  readonly " $fragmentType": "TodoListFooter_user";
 };
-export type TodoListFooter_user$data = TodoListFooter_user;
+export type TodoListFooter_user = TodoListFooter_user$data;
 export type TodoListFooter_user$key = {
-    readonly " $data"?: TodoListFooter_user$data;
-    readonly " $fragmentRefs": FragmentRefs<"TodoListFooter_user">;
+  readonly " $data"?: TodoListFooter_user$data;
+  readonly " $fragmentSpreads": FragmentRefs<"TodoListFooter_user">;
 };
-
-
 
 const node: ReaderFragment = (function(){
 var v0 = {
@@ -159,5 +163,7 @@ return {
   "abstractKey": null
 };
 })();
-(node as any).hash = '3225adf0f3b5d0a46ee4c63c8dbd9b08';
+
+(node as any).hash = "3225adf0f3b5d0a46ee4c63c8dbd9b08";
+
 export default node;
