@@ -1,33 +1,37 @@
+/**
+ * @generated SignedSource<<307064e528a75fcd73eae513cb480faf>>
+ * @lightSyntaxTransform
+ * @nogrep
+ */
+
 /* tslint:disable */
 /* eslint-disable */
 // @ts-nocheck
 
-import { ReaderFragment } from "relay-runtime";
+import { Fragment, ReaderFragment } from 'relay-runtime';
 import { FragmentRefs } from "relay-runtime";
-export type TodoList_user = {
-    readonly todos: {
-        readonly edges: ReadonlyArray<{
-            readonly node: {
-                readonly id: string;
-                readonly complete: boolean;
-                readonly " $fragmentRefs": FragmentRefs<"Todo_todo">;
-            } | null;
-        } | null> | null;
-    } | null;
-    readonly id: string;
-    readonly userId: string;
-    readonly totalCount: number;
-    readonly completedCount: number;
-    readonly " $fragmentRefs": FragmentRefs<"Todo_user">;
-    readonly " $refType": "TodoList_user";
+export type TodoList_user$data = {
+  readonly todos: {
+    readonly edges: ReadonlyArray<{
+      readonly node: {
+        readonly id: string;
+        readonly complete: boolean;
+        readonly " $fragmentSpreads": FragmentRefs<"Todo_todo">;
+      } | null;
+    } | null> | null;
+  } | null;
+  readonly id: string;
+  readonly userId: string;
+  readonly totalCount: number;
+  readonly completedCount: number;
+  readonly " $fragmentSpreads": FragmentRefs<"Todo_user">;
+  readonly " $fragmentType": "TodoList_user";
 };
-export type TodoList_user$data = TodoList_user;
+export type TodoList_user = TodoList_user$data;
 export type TodoList_user$key = {
-    readonly " $data"?: TodoList_user$data;
-    readonly " $fragmentRefs": FragmentRefs<"TodoList_user">;
+  readonly " $data"?: TodoList_user$data;
+  readonly " $fragmentSpreads": FragmentRefs<"TodoList_user">;
 };
-
-
 
 const node: ReaderFragment = (function(){
 var v0 = {
@@ -87,16 +91,16 @@ return {
                   "storageKey": null
                 },
                 {
+                  "args": null,
+                  "kind": "FragmentSpread",
+                  "name": "Todo_todo"
+                },
+                {
                   "alias": null,
                   "args": null,
                   "kind": "ScalarField",
                   "name": "__typename",
                   "storageKey": null
-                },
-                {
-                  "args": null,
-                  "kind": "FragmentSpread",
-                  "name": "Todo_todo"
                 }
               ],
               "storageKey": null
@@ -171,5 +175,7 @@ return {
   "abstractKey": null
 };
 })();
-(node as any).hash = '4c169798c328a2b4d9b4ae5227d016bb';
+
+(node as any).hash = "4c169798c328a2b4d9b4ae5227d016bb";
+
 export default node;
