@@ -1,46 +1,37 @@
+/**
+ * @generated SignedSource<<7e2f1da5b281b8f25abd80d32305490b>>
+ * @lightSyntaxTransform
+ * @nogrep
+ */
+
 /* tslint:disable */
 /* eslint-disable */
 // @ts-nocheck
 
-import { ConcreteRequest } from "relay-runtime";
+import { ConcreteRequest, Mutation } from 'relay-runtime';
 export type RemoveTodoInput = {
-    id: string;
-    userId: string;
-    clientMutationId?: string | null;
+  id: string;
+  userId: string;
+  clientMutationId?: string | null;
 };
-export type useRemoveTodoMutationVariables = {
-    input: RemoveTodoInput;
+export type useRemoveTodoMutation$variables = {
+  input: RemoveTodoInput;
 };
-export type useRemoveTodoMutationResponse = {
-    readonly removeTodo: {
-        readonly deletedTodoId: string;
-        readonly user: {
-            readonly completedCount: number;
-            readonly totalCount: number;
-        };
-    } | null;
+export type useRemoveTodoMutationVariables = useRemoveTodoMutation$variables;
+export type useRemoveTodoMutation$data = {
+  readonly removeTodo: {
+    readonly deletedTodoId: string;
+    readonly user: {
+      readonly completedCount: number;
+      readonly totalCount: number;
+    };
+  } | null;
 };
+export type useRemoveTodoMutationResponse = useRemoveTodoMutation$data;
 export type useRemoveTodoMutation = {
-    readonly response: useRemoveTodoMutationResponse;
-    readonly variables: useRemoveTodoMutationVariables;
+  variables: useRemoveTodoMutationVariables;
+  response: useRemoveTodoMutation$data;
 };
-
-
-
-/*
-mutation useRemoveTodoMutation(
-  $input: RemoveTodoInput!
-) {
-  removeTodo(input: $input) {
-    deletedTodoId
-    user {
-      completedCount
-      totalCount
-      id
-    }
-  }
-}
-*/
 
 const node: ConcreteRequest = (function(){
 var v0 = [
@@ -164,5 +155,7 @@ return {
   }
 };
 })();
-(node as any).hash = 'fd8823db59682a3723699ef8ffac4f83';
+
+(node as any).hash = "fd8823db59682a3723699ef8ffac4f83";
+
 export default node;
