@@ -1,43 +1,36 @@
+/**
+ * @generated SignedSource<<bf0a6e6e13c56d8f9f563a2bf95a3292>>
+ * @lightSyntaxTransform
+ * @nogrep
+ */
+
 /* tslint:disable */
 /* eslint-disable */
 // @ts-nocheck
 
-import { ConcreteRequest } from "relay-runtime";
+import { ConcreteRequest, Mutation } from 'relay-runtime';
 export type RenameTodoInput = {
-    id: string;
-    text: string;
-    clientMutationId?: string | null;
+  id: string;
+  text: string;
+  clientMutationId?: string | null;
 };
-export type useRenameTodoMutationVariables = {
-    input: RenameTodoInput;
+export type useRenameTodoMutation$variables = {
+  input: RenameTodoInput;
 };
-export type useRenameTodoMutationResponse = {
-    readonly renameTodo: {
-        readonly todo: {
-            readonly id: string;
-            readonly text: string;
-        };
-    } | null;
+export type useRenameTodoMutationVariables = useRenameTodoMutation$variables;
+export type useRenameTodoMutation$data = {
+  readonly renameTodo: {
+    readonly todo: {
+      readonly id: string;
+      readonly text: string;
+    };
+  } | null;
 };
+export type useRenameTodoMutationResponse = useRenameTodoMutation$data;
 export type useRenameTodoMutation = {
-    readonly response: useRenameTodoMutationResponse;
-    readonly variables: useRenameTodoMutationVariables;
+  variables: useRenameTodoMutationVariables;
+  response: useRenameTodoMutation$data;
 };
-
-
-
-/*
-mutation useRenameTodoMutation(
-  $input: RenameTodoInput!
-) {
-  renameTodo(input: $input) {
-    todo {
-      id
-      text
-    }
-  }
-}
-*/
 
 const node: ConcreteRequest = (function(){
 var v0 = [
@@ -118,5 +111,7 @@ return {
   }
 };
 })();
-(node as any).hash = 'bec01769317f1331bc83080e3dee1883';
+
+(node as any).hash = "bec01769317f1331bc83080e3dee1883";
+
 export default node;
